new stack/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
faster-whisper
pyannote.audio
soundfile
torch
torchvision
torchaudio
speechbrain
omegaconf
einops
onnxruntime-gpu
python-dotenv
